--- app/views/users/page.html.slim
@@ -1,4 +1,0 @@
-= javascript_include_tag 'app/user_pages', 'data-turbolinks-track': 'reload'
-
-script
-  == render partial: controller.action_name, formats: :js
